Changement noms méthodes, amélioration qualité code

+++ ERP/Mobile.pde
@@ -1,8 +1,8 @@
 public class Mobile {
     private float x,y;
     private float x0, y0;
-    private int vx, vy;
-    public Mobile(float x0, float y0, int vx, int vy) {
+    private float vx, vy;
+    public Mobile(float x0, float y0, float vx, float vy) {
         this.x=x0;
         this.y=y0;
         this.x0=x0;
@@ -27,10 +27,10 @@ public class Mobile {
     public float getY() {
         return y;
     }
-    public int getVx() {
+    public float getVx() {
         return vx;
     }
-    public int getVy() {
+    public float getVy() {
         return vy;
     }
 }
